Agrega componente para nuevo pedido.
