--- src/components/NewOrder.jsx
@@ -1,5 +1,0 @@
-const NewOrder = () => {
-	return <div></div>;
-};
-
-export default NewOrder;
